LayoutTests/canvas/philip/tests/2d.imageData.put.wrongtype-expected.txt

< [index] >
2d.imageData.put.wrongtype
putImageData() does not accept non-ImageData objects
References: 2d.imageData.put.wrongtype
Actual output:
Expected output:

Passed
